review of mouse_test
